przeslanie leku z magazynu na sale, w wyszukiwaniu zawezenie do magazynu

+++ szs/src/main/java/psk/pip/project/szs/dto/administration/DrugTransferDTO.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class DrugTransferDTO {
 	private Long roomId;
 	private Long drugId;
-	private Long drugsAmountToMove;
+	private Integer drugsAmountToMove;
 }
